+++ mkl_umath/__init__.py
@@ -29,7 +29,14 @@ Intel(R) C compiler runtime.
 """
 
 from . import _init_helper
-from ._patch import is_patched, mkl_umath, restore, use_in_numpy
+from ._patch_numpy import (
+    is_patched,
+    mkl_umath,
+    patch_numpy_umath,
+    restore,
+    restore_numpy_umath,
+    use_in_numpy,
+)
 from ._ufuncs import *
 from ._version import __version__